Extract general components to new file
