+++ src/components/Statistics.tsx
@@ -40,7 +40,8 @@ import {Legend} from "@visx/visx";
 import {useLiveQuery} from "dexie-react-hooks";
 import NoData from "./NoData";
 import {Loading} from "./Loading";
-import {LayoutContext, Main, Title} from "../App";
+import {LayoutContext, Main} from "../App";
+import {Title} from "./GeneralComponents";
 
 const width = 200, height = 200;
 const radius = Math.min(width, height) / 2;
